refactor: move the video-model cache verify/repair engine into a service (#5711)
The HuggingFace cache engine — enumerating every weight, text-encoder and
IC-LoRA repo a video model needs, keying them, running shallow/deep integrity
verification and driving repair — lived between route registrations in
server/routes/videoGen.js despite never touching req/res. It now sits in
server/services/videoGen/modelCache.js beside runtimes.js and
runtimeInstaller.js, which answer the same "is this install ready to render"
question, so the media job queue and maintenance scripts can reach it without
importing an Express router.

The route file drops 169 lines and keeps only HTTP glue; every affected
endpoint (/models/status, /models/verify, /models/:modelId/repair, and the
IC-LoRA and text-encoder download/repair pairs) keeps its exact handler body
and payload.

Deviation from the filed plan: textEncoderDownloadTarget is exported rather
than kept module-private, because the /text-encoders/:id download and repair
routes call it directly.
